requirements.txt
flask==3.0.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
xgboost>=2.0.0
requests>=2.31.0
joblib>=1.3.0
gunicorn==21.2.0
pyyaml>=6.0
openpyxl>=3.1.0
typing-extensions>=4.5.0
nba_api>=1.2.0
